Radu

+++ PolyCook/View/RecipeListView.swift
@@ -3,23 +3,7 @@ import SwiftUI
 struct RecipeListView: View {
     @EnvironmentObject var loginVM : LoginViewModel
     
-    @State var recipes = ["Steak frites", "Mousse au chocolat", "Boeuf bourgignon", "Ratatouille", "Pates carbonara", "Omelette", "Oeufs mimosa", "Tarte thon", "Coquilles St. Jacques", "Velouté de potiron", "Soupe à l'oignon", "Salade César", "Oeufs cocottes", "Quiche aux poireaux", "Nems de figatelli", "Bruschetta"]
-    @State private var enteredText : String = ""
-    
-    @State var toBeDeleted : IndexSet?
-    @State var showingDeleteAlert = false
-    
-    //filter states
-    @State var showMealFilter = false
-    @State var showIngredientFilter = false
-    @State var mealFilters = [FilterItem(title: "Entrée"), FilterItem(title: "Principal"), FilterItem(title: "Dessert")]
-    @State var ingredientFilters = [FilterItem(title: "Pomme"), FilterItem(title: "Oeuf"), FilterItem(title: "Pâtes"), FilterItem(title: "Poisson"), FilterItem(title: "Tomate"), FilterItem(title: "Oignon"), FilterItem(title: "Courgette")]
-    
-    
-    func showConfirmation(at indexSet : IndexSet) {
-        self.toBeDeleted = indexSet
-        self.showingDeleteAlert = true
-    }
+    @ObservedObject var listVm = RecipeListViewModel()
     
     var body: some View {
         NavigationView{
@@ -29,7 +13,7 @@ struct RecipeListView: View {
                         Section {
                             HStack(spacing: 15) {
                                 Button (action: {
-                                    withAnimation{showMealFilter.toggle()}
+                                    withAnimation{listVm.showMealFilter.toggle()}
                                 }, label: {
                                     Text("Type Repas").font(.system(size: 21))
                                 })
@@ -40,7 +24,7 @@ struct RecipeListView: View {
                                     .buttonStyle(BorderlessButtonStyle())
                                 
                                 Button (action: {
-                                    withAnimation{showIngredientFilter.toggle()}
+                                    withAnimation{listVm.showIngredientFilter.toggle()}
                                 }, label: {
                                     Text("Ingrédients").font(.system(size: 21))
                                 })
@@ -55,46 +39,58 @@ struct RecipeListView: View {
                         .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
                         
                         Section {
-                            ForEach(searchResults, id: \.self) {recipe in
-                                ZStack {
-                                    NavigationLink(destination: RecipeView(isSheet: false)) {
-                                        EmptyView()
-                                    }.opacity(0)
-                                    HStack {
-                                        Text(recipe).font(.system(size: 21)).truncationMode(.tail)
-                                        Spacer()
-                                        Image(systemName: "exclamationmark.circle")
-                                            .resizable()
-                                            .frame(width: 25, height: 25)
-                                            .foregroundColor(.red)
-                                    }.frame(height: 50)
-                                }
+                            if searchResults.isEmpty {
+                                Text("Aucune recette")
+                                    .font(.system(size: 21))
+                                    .bold()
+                                    .frame(maxWidth: .infinity)
+                                    .padding(.vertical, 20)
+                                    .listRowBackground(Color.white.opacity(0))
+                                    .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
                             }
-                            .onDelete(perform: showConfirmation)
-                            .confirmationDialog(
-                                "Voulez vous supprimer la recette?",
-                                isPresented : $showingDeleteAlert,
-                                titleVisibility: .visible
-                            ) {
-                                Button("Oui") {
-                                    withAnimation {
-                                        for index in self.toBeDeleted! {
-                                            recipes.remove(at: index)
-                                        }
+                            else {
+                                ForEach(searchResults, id: \.id) {recette in
+                                    ZStack {
+                                        NavigationLink(destination: RecipeView(recette: recette, isSheet: false)) {
+                                            EmptyView()
+                                        }.opacity(0)
+                                        HStack {
+                                            Text(recette.nomRecette).font(.system(size: 21)).truncationMode(.tail)
+    //                                        Spacer()
+    //                                        Image(systemName: "exclamationmark.circle")
+    //                                            .resizable()
+    //                                            .frame(width: 25, height: 25)
+    //                                            .foregroundColor(.red)
+                                        }.frame(height: 50)
                                     }
                                 }
-                                Button("Non", role: .cancel) {}
+                                .onDelete(perform: listVm.showConfirmation)
+                                .confirmationDialog(
+                                    "Voulez vous supprimer la recette?",
+                                    isPresented : $listVm.showingDeleteAlert,
+                                    titleVisibility: .visible
+                                ) {
+                                    Button("Oui") {
+                                        withAnimation {
+                                            for index in self.listVm.toBeDeleted! {
+                                                listVm.recipes.remove(at: index)
+                                            }
+                                        }
+                                    }
+                                    Button("Non", role: .cancel) {}
+                                }
+                                .deleteDisabled(!loginVM.isSignedIn)
                             }
-                            .deleteDisabled(!loginVM.isSignedIn)
+                            
                         }
                     }
-                    .searchable(text: $enteredText, placement: .navigationBarDrawer(displayMode: .always), prompt: "Recherche recette")
+                    .searchable(text: $listVm.enteredText, placement: .navigationBarDrawer(displayMode: .always), prompt: "Recherche recette")
                     .navigationTitle("Recettes")
                     .toolbar {
                         ToolbarItem(placement: .navigationBarTrailing) {
                             if loginVM.signedIn {
                                 Button{} label: {
-                                    NavigationLink(destination: CreateRecipeView()){
+                                    NavigationLink(destination: CreateRecipeView(listVm : self.listVm)){
                                         Image(systemName: "text.badge.plus")
                                     }
                                 }
@@ -102,19 +98,19 @@ struct RecipeListView: View {
                         }
                     }
                 }
-                FilterMenu(title: "Type Repas", height: 125, isOn: $showMealFilter, filters: $mealFilters)
-                FilterMenu(title: "Ingrédients", height: 250, isOn: $showIngredientFilter, filters: $ingredientFilters)
+                FilterMenu(title: "Type Repas", height: 125, isOn: $listVm.showMealFilter, filters: $listVm.mealFilters)
+                FilterMenu(title: "Ingrédients", height: 250, isOn: $listVm.showIngredientFilter, filters: $listVm.ingredientFilters)
             }
         }
         .navigationViewStyle(StackNavigationViewStyle()) //to fix constraints error that appear in the console due to navigationTitle
     }
     
-    var searchResults: [String] {
-        if enteredText.isEmpty {
-            return recipes
+    var searchResults: [Recette] {
+        if listVm.enteredText.isEmpty {
+            return listVm.recipes
         } else {
             //we need to filter using lowercased names
-            return recipes.filter { $0.lowercased().contains(enteredText.lowercased())}
+            return listVm.recipes.filter { $0.nomRecette.lowercased().contains(listVm.enteredText.lowercased())}
         }
     }
 }
